Abstract bot instance to remove nullness
